requirements.txt
absl-py==2.3.1
accelerate==1.11.0
addict==2.4.0
basicsr==1.4.2
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
diffusers==0.35.2
facexlib==0.3.0
filelock==3.20.0
filterpy==1.4.5
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
realesrgan==0.3.0
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
xformers==0.0.33.post1
yapf==0.43.0
zipp==3.23.0
